+++ Main/Features/Encyclopedia/EncyclopediaIoC.cs
@@ -1,0 +1,11 @@
+using DryIoc;
+
+namespace TAOM.Features.Encyclopedia;
+
+public static class EncyclopediaIoC
+{
+    public static void RegisterEncyclopediaFeature(IContainer container)
+    {
+        container.Register<IEncyclopediaSettingsProvider, EncyclopediaSettingsProvider>(Reuse.Singleton);
+    }
+}